mm-video: venc: handle zero-length invalid EOS buffers
Some clients (like GraphicsBufferSource) can queue meta buffer
with invalid handle for EOS, with filled length being empty.
Handle this case by queueing an internal color-conversion
buffer (if available) to propagate the EOS to output.
If a conversion buffer is not available, wait for a buffer
to be returned by the encoder.

Change-Id: Ib30e7063960939e7b764088afe9715f3c3d86d9f