Fix stride output Venus's YUV420PackedSemiPlanar.
Particularly, QOMX_COLOR_FORMATYUV420PackedSemiPlanar32m and
QOMX_COLOR_FORMATYUV420PackedSemiPlanar32mMultiView. Recently, I tested
generating thumbnail for 480x320 video, and found that it was not
working. So, I goes on finding what's the actual stride and
slice-height alignment for this format.

So, I read file include/media/msm_media_info.h in kernel tree, and it
states clearly that the alignment for stride is 128 while the alignment
for silce-height 32. This commit reflect this actual value, and fix
thumbnail generation for 480x320 video on Ubuntu Touch.